+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - discogs: Show credits hint on the Discogs edit page
 - popup: Receive bandcamp artist schema
 - popup: Show apply metadata icon only on Discogs edit page
+- core: Use a generic method for sending messages from popup to the page
 
 ## 0.12.0 (2024-08-16)